+++ src/communities/communities.route.js
@@ -1,12 +1,16 @@
 import express from 'express';
 import asyncHandler from 'express-async-handler';
-import { createCommunityController } from './communities.controllers.js';
-import { communityUsersRouter } from '../communityUsers/communityUsers.route.js';
-import { getCommunitiesController } from './communities.controllers.js';
+import { joinCommunityController } from './communities.controllers.js';
 
+// 라우터 인스턴스 생성
 export const communitiesRouter = express.Router();
-communitiesRouter.get('/', asyncHandler(getCommunitiesController));
-communitiesRouter.post('/create', asyncHandler(createCommunityController));
-communitiesRouter.post('/join', communityUsersRouter); // 커뮤니티 사용자 관련 라우터 추가
 
+// 사용자 ID를 직접 설정하는 미들웨어 (이렇게 설정하면 모든 요청에 동일한 사용자 ID가 적용됩니다)
+const setUserMiddleware = (req, res, next) => {
+    req.user = { id: 5 }; // 임시 사용자 ID: 6으로 설정
+    next();
+};
+
+// 커뮤니티 가입 엔드포인트
+communitiesRouter.post('/join', setUserMiddleware, asyncHandler(joinCommunityController));
